+++ Model/StatisticProvider/BooleanProvider.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+namespace Alekseon\CustomFormsStatistics\Model\StatisticProvider;
+
+/**
+ *
+ */
+class BooleanProvider extends  DefaultProvider
+{
+
+}