+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 
 ## [Unreleased]
+### Fixed
+- fix for statistic filters
 
 ## [100.0.2] - 2023-04-19
 ### Changed